Rename major,minor,patch to update,info,trace
Also add a timestamp to appmaps and fingerprints, to avoid
re-fingerprinting
