+++ src/lib/fingerprint/canonicalize/info_v1.js
