+++ src/lib/fingerprint/canonicalize/trace_v1.js
