+++ src/lib/fingerprint/canonicalize/update_v1.js
